requirements.txt
GitPython
commentjson
gevent
steam[client]
eventemitter
python-dotenv